rename archive commands to `archive dump` and `archive load`
