+++ src/client/cli/commands/archive/load.js
@@ -3,20 +3,21 @@
 const fs = require('fs')
 const gunzip = require('gunzip-maybe')
 const tar = require('tar-stream')
-const {consumeStream} = require('../../../common/util')
-const {subcommand, printlnErr, println, pluralizeCount} = require('../util')
+const {consumeStream} = require('../../../../common/util')
+const {subcommand, printlnErr, println, pluralizeCount} = require('../../util')
 
-import type {RestClient} from '../../api'
+import type {RestClient} from '../../../api'
 import type {Readable as ReadableStream} from 'stream'
 
 const OBJECT_BATCH_SIZE = 1024
 
 module.exports = {
-  command: 'archiveLoad [filename]',
+  command: 'load [filename]',
+  description: 'Load an archive (generated by `mcclient archive dump`) into the local node.\n',
   builder: {
     filename: {
       description: 'A tar archive (optionally gzipped) from which to read statements and objects, ' +
-        'as generated by `mcclient archive`.  If not given, archive will be read from standard input.\n',
+        'as generated by `mcclient archive dump`.  If not given, archive will be read from standard input.\n',
       required: false,
       type: 'string',
       default: null
